Add '--root' option in text2hdf5 to load histograms with ROOT, otherwise uproot
